# Improve performance of search page