--- a/app/views/channels/_search_result_row.html.erb
+++ b/app/views/channels/_search_result_row.html.erb
@@ -1,0 +1,61 @@
+<tr>
+  <td>
+    <% if channel.thumbnail_url.present? %>
+      <%= link_to(image_tag(channel.thumbnail_url), channel_path(channel, take_params)) %>
+    <% end %>
+  </td>
+  <td>
+    <% if channel.disabled? %>
+      <span data-toggle="tooltip" title="<%= t('helpers.notice.channel_disabled') %>">⚠️</span>
+    <% end %>
+    <%= print_link(channel.title, channel_path(channel, take_params), t('text.channel.title.missing')) %>
+    <div>
+      <% channel.tags.each do |tag| %>
+        <%= link_to(tag.name, channels_path(search_channel: { tag: tag.name }), class: 'badge rounded-pill bg-secondary', remote: true) %>
+      <% end %>
+    </div>
+  </td>
+  <td class="d-none d-md-table-cell">
+    <%= tag.span print_datetime(channel.published_at, format: :date) %>
+  </td>
+  <td>
+    <div class="vertical text-end text-nowrap">
+      <span>
+        <%= print_number(channel.subscriber_count) %>
+      </span>
+      <span>
+        <%= print_diff_numbers(channel.subscriber_count, channel.second_latest_subscriber_count) %>
+      </span>
+    </div>
+  </td>
+  <td class="d-none d-md-table-cell">
+    <div class="vertical text-end text-nowrap">
+      <span>
+        <%= print_number(channel.view_count) %>
+      </span>
+      <span>
+        <%= print_diff_numbers(channel.view_count, channel.second_latest_view_count) %>
+      </span>
+    </div>
+  </td>
+  <td class="d-none d-md-table-cell">
+    <div class="vertical text-end text-nowrap">
+      <span>
+        <%= print_number(channel.video_count) %>
+      </span>
+      <span>
+        <%= print_diff_numbers(channel.video_count, channel.second_latest_video_count) %>
+      </span>
+    </div>
+  </td>
+  <td class="d-none d-xl-table-cell">
+    <div class="vertical text-end text-nowrap">
+      <span>
+        <%= print_acquired_at(channel.latest_acquired_at, format: :date) %>
+      </span>
+      <span>
+        <%= print_comparison_period(channel.latest_acquired_at, channel.second_latest_acquired_at) %>
+      </span>
+    </div>
+  </td>
+</tr>