Rendering of search results is taking a long time. The log shows that the value of Allocations is very large.
```
Completed 200 OK in 1262ms (Views: 1230.7ms | ActiveRecord: 15.0ms | Allocations: 1467587)
```

Change the search logic to improve performance.

Log after improvement:
```
Completed 200 OK in 221ms (Views: 157.2ms | ActiveRecord: 54.9ms | Allocations: 171880)
```

Details are below:
- Stop preloading channel_statistics since it does not need to read all of them.
- This has nothing to do with performance improvement: Fix the search to include channel records with no channel_statistics.

Reference:
- [Rails(ActiveRecord)で自在にSQLを書く手段いろいろ](https://takakisan.com/rails-sql/)
- [【SQL】グループごとに最大の値を持つレコードを取得する方法3選](https://takakisan.com/sql-max-in-each-group/)